Removed BiFunction, added hot/warm/cold path indicator in EPP, changed comment in HSE.
BiFunction was removed because this package/library should be used for
/annotations/ only and should avoid shipping interfaces and/or utility
classes. There is a place for that code, but this library isn't one.

Hot/warm/cold path indicator in Expected performance profile is optional
and is backward compatible (old implementations don't have to change). I'm
not 110% sure if this is the best place for it, but there's no real point
in creating another annotation related to performance.

Comment in HasSideEffects was misleading - it wasn't optional, so it had
to be changed.